+++ src/parser/index.ts
@@ -12,7 +12,7 @@ import { parseSectionImport } from './import';
 import { parseSectionType } from './type';
 import { parseSectionCanon } from './canon';
 import { parseSectionCoreInstance } from './coreInstance';
-import { parseComponent } from './component';
+import { parseSectionInstance } from './instance';
 
 export const WIT_MAGIC = [0x00, 0x61, 0x73, 0x6d];
 export const WIT_VERSION = [0x0D, 0x00];
@@ -95,7 +95,7 @@ async function parseSection(ctx: ParserContext, src: Source): Promise<WITSection
             case 0: return parseSectionCustom(ctx, sub!, size);
             case 1: return parseModule(ctx, asyncSub!, size);
             case 2: return parseSectionCoreInstance(ctx, sub!);
-            case 4: return parseComponent(ctx, sub!); // component
+            case 5: return parseSectionInstance(ctx, sub!);
             case 6: return parseSectionAlias(ctx, sub!);
             case 7: return parseSectionType(ctx, sub!);
             case 8: return parseSectionCanon(ctx, sub!);
@@ -104,7 +104,7 @@ async function parseSection(ctx: ParserContext, src: Source): Promise<WITSection
 
             //TODO: to implement
             case 3: // core type - we don't have it in the sample
-            case 5: // instance
+            case 4: // component
                 return skipSection(ctx, sub!, type, size); // this is all TODO
             default:
                 throw new Error(`unknown section: ${type}`);
